FF-48: lavet alle 3 grafer i toppen, de loader stadivæk ikke ind med det samme

+++ lib/models/graph_data_model.dart
@@ -1,6 +1,6 @@
 class GraphData {
   int month;
-  int maxWeight;
+  double maxWeight;
   int minWeight;
   int trainingCount;
